Rename model lib to printing-engine
