+++ src/app/lib/polling.js
@@ -1,13 +1,13 @@
 // @flow
 
 import type {ModelId} from '../type-next'
-import * as modelLib from './model'
+import * as printingEngine from './printing-engine'
 
 export const POLLING_FAILED = Symbol('POLLING_FAILED')
 
 export const pollingFunction = {
   modelSceneId: async (modelId: ModelId) => {
-    const {model} = await modelLib.getModelWithStatus(modelId)
+    const {model} = await printingEngine.getModelWithStatus(modelId)
 
     return typeof model.sceneId === 'string' ? model.sceneId : POLLING_FAILED
   }
